created components folder and moved files, moved app.scss to partials, and this includes Alexs latest App.js
