--- src/App.scss
@@ -1,4 +1,0 @@
-@import "./styles/setup";
-@import "./styles/partials";
-@import "./styles/global";
-@import "./styles/header";
