+++ src/components/SavedGame.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import firebase from './firebase.js';
+import firebase from '../firebase.js';
 
 class SavedGame extends Component {
   constructor(props) {
