+++ src/styles/App.scss
@@ -1,0 +1,4 @@
+@import 'setup';
+@import 'partials';
+@import 'global';
+@import 'header';
